[BEAM-10602] Add latency/checkpoint duration as separate panels in ParDo Load Test
The Flink streaming tests were reported in a separate table and made avaible
through this dashboard: https://apache-beam-testing.appspot.com/explore?dashboard=5751884853805056

Turns out, this is not optimal for the new Grafana-based dashboard. We have to
change the table name because the query capability of InfluxDb is very limited.

This way the results will be shown together with the other Runners' load test results.